// src/api/models/Definition.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Definition = {
    readonly id: number;
    translation: string;
    part_of_speech: string;
    details: string;
    example: string;
    direct_synonyms: string;
};